Add missing casts

+++ include/pgm/pgm_index.hpp
@@ -252,7 +252,7 @@ struct PGMIndex<K, Epsilon, EpsilonRecursive, Floating>::Segment {
      * @return the approximate position of the specified key
      */
     inline size_t operator()(const K &k) const {
-        auto pos = int64_t(slope * (k - key)) + intercept;
+        auto pos = int64_t(slope * double(k - key)) + intercept;
         return pos > 0 ? size_t(pos) : 0ull;
     }
 };
